Refactoring in preparation for dateline algorithm implementation.
Starting to generalize address and DOR for arbitrary dimensions.
Added a "Coordinate" typedef.  Changed wrap point such that it's now a Coordinate and therefore unsigned.
I think I can get away with using zero for NO_WRAP since it would be a very weird topology with a single
node that had 4 links to itself, 2 of which were considered "wrap links".
Added a helper function which converts a 3d torus port number into a string like "y-".