Bugfix in TocinoRx::Receive and associated cleanup.
During enqueue we need to detect the transition between unblocked and blocked.
It is only in this case where we want to pause the remote node.  Previously,
the pre-enqueue check was queue->IsAlmostFull, which is wrong, and could lead
to us "double pausing."

Consider a case in which a receiver becomes blocked due to an almost full
condition on queue A, and issues a remote pause.  Then, before the remote node
can receive and process the pause, another flit arrives which causes an almost
full condition in queue B.  In this case we were previously sending another
redundant pause.  Now that the pre-enqueue check is IsVCBlocked, we will scan
across all the queues and notice that queue A was already full, and that we
were already blocked.  Thus we will not issue the remote pause.